[LAS-12] feat: 헤더 컴포넌트 적용 #181
