--- src/components/Search/Header/index.tsx
@@ -1,21 +1,0 @@
-import React from 'react';
-import * as S from './style';
-import LeftArrowIcon from '@assets/icon/left-arrow.svg';
-import { useRouter } from 'next/router';
-
-const Header = () => {
-  const router = useRouter();
-
-  const handleRouteBack = () => {
-    router.back();
-  };
-
-  return (
-    <S.Wrapper>
-      <LeftArrowIcon onClick={handleRouteBack} />
-      <span>검색</span>
-    </S.Wrapper>
-  );
-};
-
-export default Header;
