+++ src/components/SearchGym/Layout/index.tsx
@@ -1,15 +1,14 @@
-import React, { ReactNode } from 'react';
-import { Header } from '@components/SearchGym';
 import * as S from './style';
+import { Header } from '@components/Common';
 
 interface Props {
-  children: ReactNode;
+  children: React.ReactNode;
 }
 
 const Layout = ({ children }: Props) => {
   return (
     <>
-      <Header />
+      <Header text="헬스장 검색" />
       <S.Wrapper>{children}</S.Wrapper>
     </>
   );
