+++ src/pages/mypage/partners/index.tsx
@@ -1,20 +1,20 @@
-import { Content, Header } from '@components/MyPage/PartnerList';
+import { Content } from '@components/MyPage/PartnerList';
 import { useGetPartnerList } from '@hooks/MyPage/queries';
 import { Navigation } from '@components/Common';
-import { Loader } from '@components/Common';
+import { Header, Loader } from '@components/Common';
 import { useEffect, useState } from 'react';
 
 const PartnerList = () => {
   const { data: partnerListData, isError, refetch } = useGetPartnerList();
   const [isPossibleToRendering, setIsPossibleToRendering] = useState(false);
-  
+
   useEffect(() => {
     if (isError) {
       setIsPossibleToRendering(true);
     }
     if (partnerListData) {
       setIsPossibleToRendering(true);
-    } 
+    }
     if (!isPossibleToRendering) {
       refetch();
     }
@@ -28,7 +28,7 @@ const PartnerList = () => {
 
   return (
     <>
-      <Header />
+      <Header text="운동 파트너 목록" />
       <Content partnerList={partners} />
       <Navigation />
     </>
